Better handle non-standard package names
Canonical publishes mainline kernels that have a non-standard
naming scheme. Specifically, they include "-unsigned" in their name,
e.g. linux-image-unsigned-4.20.0-042000rc7-generic

Previously we created the package name of the currently running
kernel image by simply appending the output of `uname -r` which is
not sufficient/exactly matching the actual package names.
With this patch we are looking for the correct version in the
(machine-readable) list of installed packages.

Additionally, we take the possible suffix into account when comparing
the versions of installed images, headers and modules to that of
the currently running one.